Split designer interface tests from data dashboard
